+++ src/main/java/gov/usgs/owi/nldi/controllers/LookupController.java
@@ -6,12 +6,16 @@ import java.util.*;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
+import gov.usgs.owi.nldi.services.ConfigurationService;
+import gov.usgs.owi.nldi.services.Navigation;
+import gov.usgs.owi.nldi.services.Parameters;
+import gov.usgs.owi.nldi.services.LogService;
+import gov.usgs.owi.nldi.services.PyGeoApiService;
+import gov.usgs.owi.nldi.services.AttributeService;
 import io.swagger.v3.oas.annotations.Hidden;
 import io.swagger.v3.oas.annotations.Operation;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
-import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RestController;
@@ -19,10 +23,6 @@ import org.springframework.web.bind.annotation.RestController;
 import gov.usgs.owi.nldi.dao.BaseDao;
 import gov.usgs.owi.nldi.dao.LookupDao;
 import gov.usgs.owi.nldi.dao.StreamingDao;
-import gov.usgs.owi.nldi.services.ConfigurationService;
-import gov.usgs.owi.nldi.services.LogService;
-import gov.usgs.owi.nldi.services.Navigation;
-import gov.usgs.owi.nldi.services.Parameters;
 import gov.usgs.owi.nldi.transform.CharacteristicMetadataTransformer;
 
 @RestController
@@ -31,8 +31,8 @@ public class LookupController extends BaseController {
     @Autowired
     public LookupController(LookupDao inLookupDao, StreamingDao inStreamingDao,
                             Navigation inNavigation, Parameters inParameters, ConfigurationService configurationService,
-                            LogService inLogService) {
-        super(inLookupDao, inStreamingDao, inNavigation, inParameters, configurationService, inLogService);
+                            LogService inLogService, PyGeoApiService inPygeoapiService, AttributeService inAttributeService) {
+        super(inLookupDao, inStreamingDao, inNavigation, inParameters, configurationService, inLogService, inPygeoapiService, inAttributeService);
     }
 
     //swagger documentation for /lookups/{characteristicType}/characteristics endpoint