+++ src/main/java/gov/usgs/owi/nldi/controllers/NetworkController.java
@@ -12,9 +12,11 @@ import gov.usgs.owi.nldi.dao.BaseDao;
 import gov.usgs.owi.nldi.dao.LookupDao;
 import gov.usgs.owi.nldi.dao.StreamingDao;
 import gov.usgs.owi.nldi.services.ConfigurationService;
-import gov.usgs.owi.nldi.services.LogService;
 import gov.usgs.owi.nldi.services.Navigation;
 import gov.usgs.owi.nldi.services.Parameters;
+import gov.usgs.owi.nldi.services.LogService;
+import gov.usgs.owi.nldi.services.PyGeoApiService;
+import gov.usgs.owi.nldi.services.AttributeService;
 import gov.usgs.owi.nldi.swagger.model.Feature;
 import gov.usgs.owi.nldi.transform.FeatureTransformer;
 import io.swagger.v3.oas.annotations.Operation;
@@ -39,9 +41,9 @@ public class NetworkController extends BaseController {
 
 	@Autowired
 	public NetworkController(LookupDao inLookupDao, StreamingDao inStreamingDao,
-			Navigation inNavigation, Parameters inParameters, ConfigurationService configurationService,
-			LogService inLogService) {
-		super(inLookupDao, inStreamingDao, inNavigation, inParameters, configurationService, inLogService);
+							 Navigation inNavigation, Parameters inParameters, ConfigurationService configurationService,
+							 LogService inLogService, PyGeoApiService inPygeoapiService, AttributeService inAttributeService) {
+		super(inLookupDao, inStreamingDao, inNavigation, inParameters, configurationService, inLogService, inPygeoapiService, inAttributeService);
 	}
 
 	//swagger documentation for /linked-data/{featureSource}/{featureID}/navigate/{navigationMode} endpoint