+++ src/main/java/gov/usgs/owi/nldi/dao/LookupDao.java
@@ -19,6 +19,9 @@ public class LookupDao extends BaseDao {
 	public static final String SOURCE = "source";
 	public static final String SOURCE_NAME = "sourceName";
 	public static final String COMID_LAT_LON = "comidLatLon";
+	public static final String ST_DISTANCE = "distanceFromFlowline";
+	public static final String ST_CLOSEST_POINT = "closestPointOnFlowline";
+	public static final String FEATURE_LOCATION = "featureLocation";
 	
 	private static final String NS = "lookup.";
 
@@ -37,6 +40,18 @@ public class LookupDao extends BaseDao {
 
 	public List<Map<String, Object>> getList(String objectType, Map<String, Object> parameterMap) {
 		return getSqlSession().selectList(NS + objectType, parameterMap);
+	}
+
+	public float getDistanceFromFlowline(Map<String, Object> parameterMap) {
+		return getSqlSession().selectOne(NS + ST_DISTANCE, parameterMap);
+	}
+
+	public Map<String, Object> closestPointOnFlowline(Map<String, Object> parameterMap) {
+		return getSqlSession().selectOne(NS + ST_CLOSEST_POINT, parameterMap);
+	}
+
+	public Map<String, Object> getFeatureLocation(Map<String, Object> parameterMap) {
+		return getSqlSession().selectOne(NS + FEATURE_LOCATION, parameterMap);
 	}
 
 }